The great simplification
Things were getting out of hand, so this is the beginning of a major refactor. In this commit, I've removed all dependencies and started rewriting all the commands, beginning with Create. It's been renamed to New and simplified greatly.

The goal of this refactor is to keep things simple, lightweight, and functional enough to publish an initial build to NPM. After I build my first theme with it, I'll have a better idea of where things can be improved/polished.